backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.3.0
pydantic[email]
requests==2.31.0
python-dotenv==1.0.0
pyjwt==2.8.0 
beanie==1.29.0
motor==3.1.0
pymongo==4.10.1
email-validator